@@ -532,7 +532,7 @@ function strip_bbcode(&$text, $uid = '')
 
 	if (preg_match('#^<[rt][ >]#', $text))
 	{
-		$text = $phpbb_container->get('text_formatter.utils')->clean_formatting($text);
+		$text = utf8_htmlspecialchars($phpbb_container->get('text_formatter.utils')->clean_formatting($text));
 	}
 	else
 	{
@@ -1011,7 +1011,7 @@ function make_clickable($text, $server_url = false, string $class = 'postlink')
 		if (preg_match($magic_args[0], $text, $matches))
 		{
 			// Only apply $class from the corresponding function call argument (excepting emails which never has a class)
-			if ($magic_args[3] != $static_class && $magic_args[1] != MAGIC_URL_EMAIL)
+			if ($magic_args[1] != MAGIC_URL_EMAIL && $magic_args[3] != $static_class)
 			{
 				continue;
 			}
