@@ -14,13 +14,13 @@
 namespace phpbb\attachment;
 
 use phpbb\auth\auth;
-use \phpbb\cache\service;
-use \phpbb\config\config;
-use \phpbb\event\dispatcher;
-use \phpbb\language\language;
-use \phpbb\mimetype\guesser;
-use \phpbb\plupload\plupload;
-use \phpbb\user;
+use phpbb\cache\service;
+use phpbb\config\config;
+use phpbb\event\dispatcher;
+use phpbb\language\language;
+use phpbb\mimetype\guesser;
+use phpbb\plupload\plupload;
+use phpbb\user;
 
 /**
  * Attachment upload class
