@@ -21,7 +21,7 @@ class mssql_extractor extends base_extractor
 	* Writes closing line(s) to database backup
 	*
 	* @return null
-	* @throws \phpbb\db\extractor\exception\extractor_not_initialized_exception when calling this function before init_extractor()
+	* @throws extractor_not_initialized_exception when calling this function before init_extractor()
 	*/
 	public function write_end()
 	{
@@ -195,7 +195,7 @@ class mssql_extractor extends base_extractor
 	*
 	* @param	string	$table_name	name of the database table
 	* @return null
-	* @throws \phpbb\db\extractor\exception\extractor_not_initialized_exception when calling this function before init_extractor()
+	* @throws extractor_not_initialized_exception when calling this function before init_extractor()
 	*/
 	protected function write_data_mssqlnative($table_name)
 	{
@@ -311,7 +311,7 @@ class mssql_extractor extends base_extractor
 	*
 	* @param	string	$table_name	name of the database table
 	* @return null
-	* @throws \phpbb\db\extractor\exception\extractor_not_initialized_exception when calling this function before init_extractor()
+	* @throws extractor_not_initialized_exception when calling this function before init_extractor()
 	*/
 	protected function write_data_odbc($table_name)
 	{
