@@ -22,7 +22,7 @@ class installer_configuration implements ConfigurationInterface
 	/**
 	 * Generates the configuration tree builder.
 	 *
-	 * @return \Symfony\Component\Config\Definition\Builder\TreeBuilder The tree builder
+	 * @return TreeBuilder The tree builder
 	 */
 	public function getConfigTreeBuilder()
 	{
