@@ -36,7 +36,7 @@ class add_modules extends \phpbb\install\task_base
 	protected $extension_manager;
 
 	/**
-	 * @var \phpbb\install\helper\iohandler\iohandler_interface
+	 * @var iohandler_interface
 	 */
 	protected $iohandler;
 
