@@ -55,7 +55,7 @@ abstract class report_handler implements report_handler_interface
 	 *
 	 * @param \phpbb\db\driver\driver_interface	$db
 	 * @param \phpbb\event\dispatcher_interface	$dispatcher
-	 * @param \phpbb\config\db					$config
+	 * @param \phpbb\config\config				$config
 	 * @param \phpbb\auth\auth					$auth
 	 * @param \phpbb\user						$user
 	 * @param \phpbb\notification\manager		$notification
