@@ -207,7 +207,7 @@ class factory implements \phpbb\textformatter\cache_interface
 		* Modify the s9e\TextFormatter configurator before the default settings are set
 		*
 		* @event core.text_formatter_s9e_configure_before
-		* @var \s9e\TextFormatter\Configurator configurator Configurator instance
+		* @var Configurator configurator Configurator instance
 		* @since 3.2.0-a1
 		*/
 		$vars = array('configurator');
@@ -368,7 +368,7 @@ class factory implements \phpbb\textformatter\cache_interface
 		* Modify the s9e\TextFormatter configurator after the default settings are set
 		*
 		* @event core.text_formatter_s9e_configure_after
-		* @var \s9e\TextFormatter\Configurator configurator Configurator instance
+		* @var Configurator configurator Configurator instance
 		* @since 3.2.0-a1
 		*/
 		$vars = array('configurator');
@@ -446,7 +446,7 @@ class factory implements \phpbb\textformatter\cache_interface
 	/**
 	* Configure the Autolink / Autoemail plugins used to linkify text
 	*
-	* @param  \s9e\TextFormatter\Configurator $configurator
+	* @param  Configurator $configurator
 	* @return void
 	*/
 	protected function configure_autolink(Configurator $configurator)
